initial commit
This repository contains a Python script that generates random passwords. The script can be used to generate passwords of any length, and it can include a variety of different characters, including letters, numbers, and special characters. The script also allows you to specify the length of the password, as well as the number of characters to include from each category.

Features:

    1,Generates random passwords of any length.
    2, Includes a variety of different characters, including letters, numbers, and special characters


Contributing:

Contributions to this repository are welcome. Please submit all pull requests to the master branch.